requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
cycler==0.11.0
fonttools==4.37.3
idna==3.4
kiwisolver==1.4.4
matplotlib==3.5.3
numpy==1.21.6
oauthlib==3.2.2
packaging==21.3
pandas==1.3.5
Pillow==9.2.0
pmaw==2.1.3
praw==7.6.1
prawcore==2.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
scipy==1.7.3
six==1.16.0
tweepy==4.12.1
typing-extensions==4.3.0
update-checker==0.18.0
urllib3==1.26.13
websocket-client==1.4.2
XlsxWriter==3.0.3
